restore the full plot; delete unused files
